send dhcprequests for our lease at t1
In the Mirage DHCP client, take note of the RENEWAL_T1 option when
entering the BOUND state, and send a DHCPREQUEST for the obtained lease
after that time.

As we can't call Net.listen without displacing callbacks the application has
likely registered, it's not possible to hear any response to this
DHCPREQUEST.  Without some refactoring, it won't be possible
for the DHCP client here to hear a new lease, acknowledge a server's
instruction that it not continue to use the lease it originally
obtained, or move into the REBINDING state as demanded by the RFC.

This patch simply sends the DHCPREQUEST messages as the RFC specifies
in hopes that they may result in the existing lease being retained for longer.